Put subsequence implementation into its own module
